requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.27.0

# Qwen3-VL utilities
qwen-vl-utils>=0.0.8

# SAM3 (included in transformers)
# facebook/sam3 model

# Mask utilities
pycocotools>=2.0.7

# Visualization
supervision>=0.19.0
Pillow>=10.0.0

# Configuration
pyyaml>=6.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
